+++ services/api/app/main.py
@@ -77,21 +77,6 @@ def _enqueue_analysis(track_id: int) -> None:
     q = Queue("analysis", connection=_get_redis_connection())
     q.enqueue("worker.jobs.analyze_track", track_id)
 
-class TrackIn(BaseModel):
-    title: str
-    artist_name: str
-    release_title: Optional[str] = None
-    mbid: Optional[str] = None
-    duration: Optional[int] = None
-    path_local: Optional[str] = None
-
-
-class ListenIn(BaseModel):
-    user_id: str
-    played_at: datetime
-    source: Optional[str] = "listenbrainz"
-    track: TrackIn
-
 
 class TrackPathIn(BaseModel):
     path_local: str
@@ -111,104 +96,9 @@ def _week_start(dt: datetime) -> date:
     # Monday as the start of the week
     d = dt.date()
     return d - timedelta(days=d.weekday())
-
-
-def _get_or_create(db: Session, model, defaults: Dict | None = None, **kwargs):
-    inst = db.execute(select(model).filter_by(**kwargs)).scalar_one_or_none()
-    if inst:
-        return inst
-    params = {**kwargs}
-    if defaults:
-        params.update(defaults)
-    inst = model(**params)
-    db.add(inst)
-    db.flush()
-    return inst
-
 def _env(name: str, default: Optional[str] = None) -> Optional[str]:
     import os
     return os.getenv(name, default)
-
-
-def _lb_fetch_listens(user: str, since: Optional[date], token: Optional[str] = None, limit: int = 500) -> list[dict]:
-    import time
-    import requests
-
-    base = "https://api.listenbrainz.org/1/user"  # public user listens endpoint
-    params: dict = {"count": min(limit, 1000)}
-    if since:
-        # convert to unix ts at start of day
-        params["min_ts"] = int(datetime.combine(since, datetime.min.time(), tzinfo=timezone.utc).timestamp())
-
-    url = f"{base}/{user}/listens"
-    r = requests.get(url, params=params, timeout=30)
-    r.raise_for_status()
-    data = r.json()
-    return data.get("listens", [])
-
-
-def _mb_sanitize(s: Optional[str]) -> Optional[str]:
-    if s is None:
-        return None
-    return s.strip().replace("\u0000", "")
-
-
-_MB_LAST_CALL = 0.0
-
-
-def _mb_fetch_release(mbid: str) -> dict:
-    """Fetch release info from MusicBrainz with simple rate limiting."""
-    global _MB_LAST_CALL
-    wait = 1.0 - (time.time() - _MB_LAST_CALL)
-    if wait > 0:
-        time.sleep(wait)
-    _MB_LAST_CALL = time.time()
-    url = f"https://musicbrainz.org/ws/2/release/{mbid}"
-    params = {"inc": "recordings+artists", "fmt": "json"}
-    headers = {"User-Agent": "SideTrack/0.1 (+https://example.com)"}
-    r = requests.get(url, params=params, headers=headers, timeout=30)
-    r.raise_for_status()
-    return r.json()
-
-
-def _ingest_lb_rows(db: Session, listens: list[dict], user_id: str | None = None) -> int:
-    created = 0
-    for item in listens:
-        tm = item.get("track_metadata", {})
-        artist_name = _mb_sanitize(tm.get("artist_name") or tm.get("artist_name_mb")) or "Unknown"
-        track_title = _mb_sanitize(tm.get("track_name")) or "Unknown"
-        release_title = _mb_sanitize(tm.get("release_name"))
-        recording_mbid = (tm.get("mbid_mapping") or {}).get("recording_mbid")
-        played_at_ts = item.get("listened_at")
-        if not played_at_ts:
-            continue
-        played_at = datetime.utcfromtimestamp(played_at_ts)
-        uid = (user_id or item.get("user_name") or "lb").lower()
-
-        artist = _get_or_create(db, Artist, name=artist_name)
-        rel = None
-        if release_title:
-            rel = _get_or_create(db, Release, title=release_title, artist_id=artist.artist_id)
-        track = _get_or_create(
-            db,
-            Track,
-            mbid=recording_mbid,
-            title=track_title,
-            artist_id=artist.artist_id,
-            release_id=rel.release_id if rel else None,
-        )
-        exists = db.execute(
-            select(Listen).where(
-                and_(Listen.user_id == uid, Listen.track_id == track.track_id, Listen.played_at == played_at)
-            )
-        ).scalar_one_or_none()
-        if not exists:
-            db.add(Listen(user_id=uid, track_id=track.track_id, played_at=played_at, source="listenbrainz"))
-            created += 1
-    db.commit()
-    return created
-
-
 def _lastfm_fetch_tags(
     track_id: int,
     artist: str,
@@ -349,161 +239,6 @@ def health(db: Session = Depends(get_db)):
         return {"status": "ok", "db": "up"}
     except Exception as e:
         return {"status": "degraded", "db": str(e)}
-
-
-@app.post("/ingest/listens")
-def ingest_listens(
-    since: Optional[date] = Query(None),
-    listens: Optional[List[ListenIn]] = Body(None, description="List of listens to ingest"),
-    source: str = Query("auto", description="auto|listenbrainz|body|sample"),
-    db: Session = Depends(get_db),
-    user_id: str = Depends(get_current_user),
-):
-    # 3 modes: body, ListenBrainz, or sample file
-    if source == "body" and listens is None:
-        raise HTTPException(status_code=400, detail="Body listens required for source=body")
-
-    if listens is not None:
-        # Ingest provided body
-        rows = [
-            {
-                "track_metadata": {
-                    "artist_name": ls.track.artist_name,
-                    "track_name": ls.track.title,
-                    "release_name": ls.track.release_title,
-                    "mbid_mapping": {"recording_mbid": ls.track.mbid},
-                },
-                "listened_at": int(ls.played_at.timestamp()),
-                "user_name": ls.user_id,
-            }
-            for ls in listens
-            if not since or ls.played_at.date() >= since
-        ]
-        created = _ingest_lb_rows(db, rows, user_id)
-        return {"detail": "ok", "ingested": created}
-
-    if source in ("auto", "listenbrainz"):
-        token = _env("LISTENBRAINZ_TOKEN")
-        try:
-            rows = _lb_fetch_listens(user_id, since, token)
-            created = _ingest_lb_rows(db, rows, user_id)
-            return {"detail": "ok", "ingested": created, "source": "listenbrainz"}
-        except Exception as e:
-            if source == "listenbrainz":
-                raise HTTPException(status_code=502, detail=f"ListenBrainz error: {e}")
-            # fall through to sample
-
-    # sample
-    import json
-    from pathlib import Path
-    sample_path = Path("data/sample_listens.json")
-    if not sample_path.exists():
-        raise HTTPException(status_code=400, detail="No sample listens available")
-    data = json.loads(sample_path.read_text())
-    rows = []
-    for x in data:
-        dt = datetime.fromisoformat(x["played_at"].replace("Z", "+00:00"))
-        if since and dt.date() < since:
-            continue
-        rows.append(
-            {
-                "track_metadata": {
-                    "artist_name": x["track"]["artist_name"],
-                    "track_name": x["track"]["title"],
-                    "release_name": x["track"].get("release_title"),
-                    "mbid_mapping": {"recording_mbid": x["track"].get("mbid")},
-                },
-                "listened_at": int(dt.timestamp()),
-                "user_name": x["user_id"],
-            }
-        )
-    created = _ingest_lb_rows(db, rows, user_id)
-    return {"detail": "ok", "ingested": created, "source": "sample"}
-
-
-@app.post("/ingest/musicbrainz")
-def ingest_musicbrainz(release_mbid: str = Query(..., description="MusicBrainz release MBID"), db: Session = Depends(get_db)):
-    try:
-        data = _mb_fetch_release(release_mbid)
-    except requests.HTTPError as e:
-        status = getattr(e.response, "status_code", None)
-        if status == 404:
-            raise HTTPException(status_code=404, detail="release not found")
-        raise HTTPException(status_code=502, detail=f"MusicBrainz error: {e}")
-    except Exception as e:
-        raise HTTPException(status_code=502, detail=f"MusicBrainz error: {e}")
-
-    artist = None
-    ac_list = data.get("artist-credit") or []
-    if ac_list:
-        art = (ac_list[0] or {}).get("artist", {})
-        artist = _get_or_create(
-            db,
-            Artist,
-            mbid=art.get("id"),
-            defaults={"name": _mb_sanitize(art.get("name") or art.get("sort-name") or "Unknown")},
-        )
-
-    rel_date = None
-    if data.get("date"):
-        try:
-            rel_date = datetime.fromisoformat(data["date"]).date()
-        except Exception:
-            rel_date = None
-    label = None
-    labels = data.get("label-info") or data.get("label-info-list") or []
-    if labels:
-        label = _mb_sanitize((labels[0] or {}).get("label", {}).get("name"))
-
-    release = _get_or_create(
-        db,
-        Release,
-        mbid=data.get("id"),
-        defaults={
-            "title": _mb_sanitize(data.get("title") or "Unknown"),
-            "date": rel_date,
-            "label": label,
-            "artist_id": artist.artist_id if artist else None,
-        },
-    )
-
-    created_tracks = 0
-    for medium in data.get("media", []):
-        for trk in medium.get("tracks", []):
-            rec = trk.get("recording", {})
-            track_mbid = rec.get("id")
-            if not track_mbid:
-                continue
-            existing = db.execute(select(Track).filter_by(mbid=track_mbid)).scalar_one_or_none()
-            length = rec.get("length") or trk.get("length")
-            duration = int(length / 1000) if length else None
-            title = _mb_sanitize(rec.get("title") or trk.get("title") or "Unknown")
-            if existing is None:
-                db.add(
-                    Track(
-                        mbid=track_mbid,
-                        title=title,
-                        artist_id=artist.artist_id if artist else None,
-                        release_id=release.release_id,
-                        duration=duration,
-                    )
-                )
-                created_tracks += 1
-            else:
-                if artist and existing.artist_id is None:
-                    existing.artist_id = artist.artist_id
-                if existing.release_id is None:
-                    existing.release_id = release.release_id
-                if duration and existing.duration is None:
-                    existing.duration = duration
-
-    db.commit()
-    return {
-        "detail": "ok",
-        "artist_id": artist.artist_id if artist else None,
-        "release_id": release.release_id,
-        "tracks": created_tracks,
-    }
 
 
 @app.post("/tags/lastfm/sync")
@@ -714,180 +449,6 @@ def aggregate_weeks(db: Session = Depends(get_db), user_id: str = Depends(get_cu
     return {"detail": "ok", "rows": int(total)}
 
 
-@app.get("/dashboard/trajectory")
-def dashboard_trajectory(
-    window: str = Query("12w"),
-    db: Session = Depends(get_db),
-    user_id: str = Depends(get_current_user),
-):
-    # Project weekly means to 2D using (x=valence, y=energy) for simplicity
-    # window like "12w" or "6m" is parsed naively to weeks
-    try:
-        if window.endswith("w"):
-            n_weeks = int(window[:-1])
-        elif window.endswith("m"):
-            n_weeks = int(window[:-1]) * 4
-        else:
-            n_weeks = 12
-    except Exception:
-        n_weeks = 12
-
-    # collect recent weeks
-    weeks = (
-        db.execute(
-            select(MoodAggWeek.week)
-            .where(MoodAggWeek.user_id == user_id)
-            .distinct()
-            .order_by(MoodAggWeek.week.desc())
-            .limit(n_weeks)
-        ).scalars().all()
-    )
-    weeks = sorted(weeks)
-    if not weeks:
-        return {"window": window, "points": [], "arrows": []}
-
-    points = []
-    for wk in weeks:
-        axis_rows = db.execute(
-            select(MoodAggWeek.axis, MoodAggWeek.mean).where(
-                and_(MoodAggWeek.week == wk, MoodAggWeek.user_id == user_id)
-            )
-        ).all()
-        d = {ax: val for ax, val in axis_rows}
-        x = d.get("valence", 0.5)
-        y = d.get("energy", 0.5)
-        points.append({"week": str(wk), "x": x, "y": y})
-
-    arrows = []
-    for i in range(1, len(points)):
-        arrows.append({
-            "from": points[i - 1],
-            "to": points[i],
-        })
-
-    return {"window": window, "points": points, "arrows": arrows}
-
-
-@app.get("/dashboard/radar")
-def dashboard_radar(
-    week: str = Query(...),
-    db: Session = Depends(get_db),
-    user_id: str = Depends(get_current_user),
-):
-    # parse week as YYYY-MM-DD (Monday) or ISO week-like YYYY-WW (fallback)
-    wk_date: Optional[date] = None
-    try:
-        wk_date = datetime.fromisoformat(week).date()
-    except Exception:
-        try:
-            y, w = week.split("-")
-            wk_date = datetime.fromisocalendar(int(y), int(w), 1).date()
-        except Exception:
-            pass
-    if wk_date is None:
-        raise HTTPException(status_code=400, detail="Invalid week format")
-
-    rows = db.execute(
-        select(MoodAggWeek.axis, MoodAggWeek.mean).where(
-            and_(MoodAggWeek.week == wk_date, MoodAggWeek.user_id == user_id)
-        )
-    ).all()
-    axes = {ax: val for ax, val in rows}
-
-    # baseline = mean of previous 24 weeks
-    baseline_rows = db.execute(
-        select(MoodAggWeek.axis, func.avg(MoodAggWeek.mean))
-        .where(
-            and_(
-                MoodAggWeek.week < wk_date,
-                MoodAggWeek.week >= wk_date - timedelta(weeks=24),
-                MoodAggWeek.user_id == user_id,
-            )
-        )
-        .group_by(MoodAggWeek.axis)
-    ).all()
-    baseline = {ax: float(avg or 0.0) for ax, avg in baseline_rows}
-
-    # ensure all keys present
-    for ax in AXES:
-        axes.setdefault(ax, 0.0)
-        baseline.setdefault(ax, 0.0)
-
-    return {"week": str(wk_date), "axes": axes, "baseline": baseline}
-
-
-@app.get("/dashboard/outliers")
-def dashboard_outliers(
-    limit: int = Query(10, ge=1, le=100),
-    db: Session = Depends(get_db),
-    user_id: str = Depends(get_current_user),
-):
-    """Return tracks farthest from the recent mood centroid.
-
-    The centroid is computed from the user's listens over the last 90 days
-    using mood scores for each axis. Tracks are ranked by Euclidean distance
-    from this centroid.
-    """
-
-    since = datetime.now(timezone.utc) - timedelta(days=90)
-
-    # Compute centroid across recent listens
-    cent_rows = (
-        db.execute(
-            select(MoodScore.axis, func.avg(MoodScore.value))
-            .join(Listen, Listen.track_id == MoodScore.track_id)
-            .where(
-                and_(
-                    Listen.user_id == user_id,
-                    Listen.played_at >= since,
-                    MoodScore.method == DEFAULT_METHOD,
-                )
-            )
-            .group_by(MoodScore.axis)
-        ).all()
-    )
-    centroid = {ax: float(val or 0.0) for ax, val in cent_rows}
-    for ax in AXES:
-        centroid.setdefault(ax, 0.0)
-
-    # Gather candidate tracks listened to in the window
-    rows = (
-        db.execute(
-            select(Track.track_id, Track.title, Artist.name)
-            .join(Listen, Listen.track_id == Track.track_id)
-            .join(Artist, Track.artist_id == Artist.artist_id, isouter=True)
-            .where(and_(Listen.user_id == user_id, Listen.played_at >= since))
-            .group_by(Track.track_id, Track.title, Artist.name)
-        ).all()
-    )
-
-    outliers = []
-    for tid, title, artist_name in rows:
-        axis_rows = db.execute(
-            select(MoodScore.axis, MoodScore.value).where(
-                and_(
-                    MoodScore.track_id == tid,
-                    MoodScore.method == DEFAULT_METHOD,
-                )
-            )
-        ).all()
-        vec = {ax: float(val) for ax, val in axis_rows}
-        if len(vec) < len(AXES):
-            continue
-        dist = sum((vec[ax] - centroid[ax]) ** 2 for ax in AXES) ** 0.5
-        outliers.append(
-            {
-                "track_id": tid,
-                "title": title,
-                "artist": artist_name,
-                "distance": dist,
-            }
-        )
-
-    outliers.sort(key=lambda x: x["distance"], reverse=True)
-    return {"tracks": outliers[:limit], "centroid": centroid}
-
-
 @app.post("/labels")
 def submit_label(
     track_id: int,
@@ -910,3 +471,10 @@ def submit_label(
         "axis": lbl.axis,
         "value": lbl.value,
     }
+
+
+from .routes import dashboard, listens, musicbrainz
+
+app.include_router(listens.router)
+app.include_router(musicbrainz.router)
+app.include_router(dashboard.router)
